lesson6/requirements.txt
aiofiles
asyncpg
greenlet
pydantic-settings
pytz
sqlalchemy
